Add Swagger, edit Readme and config in-memory H2 database

+++ src/main/java/com/challenge/powerledger/models/BatteriesListDTO.java
@@ -8,6 +8,6 @@ import java.util.List;
 public class BatteriesListDTO {
 
     List<String> names;
-    int totalWattCapacity;
+    float totalWattCapacity;
     double averageWattCapacity;
 }
